fix(merchandising-docs): refactoring merchandising component structure to fix docs search 4 readme
