+++ packages/snap-preact-components/src/components/Atoms/Merchandising/Banner/Banner.test.tsx
@@ -4,7 +4,7 @@ import { render } from '@testing-library/preact';
 
 import { Banner } from './Banner';
 import { ContentType } from '@searchspring/snap-store-mobx';
-import { ThemeProvider } from '../../../providers';
+import { ThemeProvider } from '../../../../providers';
 
 import { MockData } from '@searchspring/snap-shared';
 import { SearchResponseModel } from '@searchspring/snapi-types';
