+++ packages/snap-preact-components/src/components/Atoms/Merchandising/Banner/index.ts
@@ -1,0 +1,1 @@
+export * from './Banner';
