+++ packages/snap-preact-components/src/components/Atoms/Merchandising/InlineBanner/index.ts
@@ -1,0 +1,1 @@
+export * from './InlineBanner';
